move login function to functions file
