+++ notfound.html
@@ -12,7 +12,6 @@
         padding: 0;
         height: 100vh;
         background: linear-gradient(to bottom right, #3494E6, #EC6EAD);
-        /* Change the colors and gradient direction as per your preference */
         display: flex;
         justify-content: center;
         align-items: center;
@@ -30,9 +29,7 @@
         height: 100vh;
         margin: 0;
         flex-direction: column;
-        /* Added flex-direction to stack items vertically */
         text-align: center;
-        /* Center text within the container */
     }
 </style>
 <div class="head">
